Install tk from conda-forge
The problem described in https://github.com/joommf/joommf/issues/81#issuecomment-342456936 can be solved by installing tk from conda-forge.

By default, the (anaconda default channel) tk is installed.

Should find a better way of making sure this happens: either in this
Dockerfile, or in the oommf conda recipe.

As a first step, I want to commit and archive this solution.

Thanks @takluyver for pointing me in the right direction.